Replace Python codegen with Haskell spec-parse-fields
The codegen Python script which parses the FIX specifications and
generates include/hffix_fields.hpp is replaced by the spec-parse-fields
Haskell program to do the same.

The new spec-parse-fields program is a better parser and picks up a few
things from the specs which were missed by the codegen script. The new
spec-parse-fields program is also a better foundation from which to further
improve FIX spec parsing.