mv spec fixspec
`git clone` will create `fixspec` and `include` directories in lexicographical
order. This will create `include/hffix_fields.hpp` file and its
`Makefile` dependencies with timestamps that indicate to `make` that
`hffix_fields.hpp` doesn't need to be rebuilt because it's newer than
its dependencies.

This way, fresh clones of hffix (either by users or Travis), to run `make test`,
will not be required to install stack and GHC and build `spec-parse-fields`.

I'm not proud of this solution, but there it is.